requirements_test.txt
pytest-homeassistant-custom-component==0.12.49
moonraker-api==2.0.5
aiohttp_cors